Pivot Lean verification to a filesystem authority machine
Building M16 showed the effort was misspent: compile_sound was near-trivial
(spec and Landlock output share a shape, so the proof was list bookkeeping),
and the whole-config differential twin mirrored a host-dependent pipeline
field-for-field, costing more to maintain than it caught. Meanwhile the bugs
that actually escape a sandbox of this kind are symlink/mount races on an
adversarially-prepared filesystem (runc's 2025 CVEs, the virtiofsd family)
and config injection, not policy-compilation errors.

Retire the twin and its bookkeeping: Model, Input, Theorems, Refinement,
Phase.lean and the whole-config effective_drt harness. Phase.lean modeled an
in-process install->run transition h5i does not have (cache-refresh/run is
phase separation across two boxes, not an fd-carrying transition).

Keep the mechanism base and the specs the product actually runs against:
- Landlock.lean gains parsePath/compileLandlock (rescued from Refinement).
- Noninterference.lean trimmed to interferesCheck + soundness + instances,
  the spec of the product-live effective::interferes; its DRT extracted to
  tests/interferes_drt.rs.
- Predict.lean, Seatbelt.lean, and the conformance probes unchanged.
- Main.lean drops DRT mode, keeps --predict/--interferes/--seatbelt.

Document the new direction in ROADMAP as section VF: an attack-driven H5iFs
authority machine (object identity, symlink/hardlink/rename, ordered mounts
with shadowing, fd table, setup state machine over an adversarial initial
state) that doubles as the semantics of a per-run translation validator on
the real policy.effective.json, with a final-state mountinfo auditor and an
optional verified-cow workspace backend. Corrects the phase-machinery claims
in V1/V3/M16.

Cost accepted: no Lean<->Rust link on the resolver until the validator lands;
that link was the maintenance burden the pivot sheds. Lean theorems and the
interferes/seatbelt DRTs are green.
