+++ lean/H5iSpec.lean
@@ -1,10 +1,5 @@
 import H5iSpec.Effective
-import H5iSpec.Input
-import H5iSpec.Model
-import H5iSpec.Theorems
 import H5iSpec.Landlock
-import H5iSpec.Phase
-import H5iSpec.Refinement
 import H5iSpec.Noninterference
 import H5iSpec.Predict
 import H5iSpec.Seatbelt
