+++ lean/H5iSpec/Predict.lean
@@ -1,4 +1,4 @@
-import H5iSpec.Refinement
+import H5iSpec.Landlock
 
 /-!
 The prediction layer for conformance probes (ROADMAP.md §V4, "model versus
