+++ lean/H5iSpec/Seatbelt.lean
@@ -1,4 +1,4 @@
-import H5iSpec.Refinement
+import H5iSpec.Landlock
 
 /-!
 The Seatbelt refinement (ROADMAP.md §V3's "each backend gets its own"): a
